requirements.txt
Flask
pymongo
python-dotenv
paho-mqtt
gunicorn